requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
openai
faiss-cpu
tiktoken
python-dotenv
